Implement product filter in list view
The product filter feature has been added to the list component. A new dependency, Reselect, has been included in the application for state selector optimization and memoization. In the redux store, a filter attribute has been introduced into the list state. The redux action responsible for updating the filter has also been created. In the list component itself, the previous selector has been replaced by a new memoized selector, productFilteredSelector, to dynamically filter listed products according to the filter state. The list updates are now based on this filtered state. A filter action has been called on the input change event in the Form component, so that whenever a user starts typing in the name field before saving, they are presented with a filtered view of their existing list of products.