Remove 'clearForm' action and refactor 'appReducer'
The 'clearForm' action was removed from the action types and interface in 'action.ts' file, due to redundancy and to enhance code efficiency. Also, initial state definition in 'appReducer' is refactored for better readability. Now each initial state property is written on a new line. This change was made to improve the readability and maintainability of the codebase.