Improve error handling and remove unused code in Form.tsx
Introduced better error handling in Form.tsx, throwing an error when there aren't any inputs in ActionTypes.FILTER and when trying to find product id. This is crucial to avoid possible runtime errors caused by null values.

Moreover, removed unused ref attributes for the "name" and "price" inputs. This change contributes towards a cleaner and optimized code as it eliminates unnecessary references.

In addition, created global-error.tsx to handle and display global errors. It will log errors in console and provide a user-friendly error message with a button to reset the application. This provides better UX when encountering unexpected situations.